requirements.txt
faster-whisper
ffmpeg-python
python-telegram-bot==22.5
huggingface-hub

